Move public data structures under OptimovePC
